auto_candidate/requirements.txt
google-generativeai
gitpython
docker
rich
typer
python-dotenv
